refacotered files, updates README and preprocessing stuff
