+++ data/curated_data/features/komodo_dragon.txt
@@ -1,0 +1,2 @@
+n_explicit=100
+n_implicit=131
