+++ data/curated_data/features/simple_komodo_dragon.txt
@@ -1,0 +1,2 @@
+n_explicit=70
+n_implicit=104
