+++ data/curated_data/simple_komodo_dragon.txt
